兼容性支持int32的message_id

+++ xposed/src/main/java/moe/fuqiuluo/http/action/data/Message.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class MessageResult(
-    @SerialName("message_id") val msgId: Long,
+    @SerialName("message_id") val msgId: Int,
     @SerialName("time") val time: Double
 )
